阿斯蒂

+++ h2/src/test/org/h2/samples/Function.java
@@ -1,6 +1,6 @@
 /*
- * Copyright 2004-2018 H2 Group. Multiple-Licensed under the MPL 2.0,
- * and the EPL 1.0 (http://h2database.com/html/license.html).
+ * Copyright 2004-2019 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
 package org.h2.samples;
@@ -145,7 +145,7 @@ public class Function {
         if (url.equals("jdbc:columnlist:connection")) {
             return rs;
         }
-        for (int s = size.intValue(), x = 0; x < s; x++) {
+        for (int s = size, x = 0; x < s; x++) {
             for (int y = 0; y < s; y++) {
                 rs.addRow(x, y);
             }
